organized types for each component
