+++ src/lib/types/index.ts
@@ -11,30 +11,3 @@ export type Coors = {
   x: number;
   y: number;
 };
-
-export type CircleProps = {
-  backgroundColor?: string;
-  color?: string;
-  clockwise?: boolean;
-  speed?: number;
-  stroke?: number;
-  radius?: number;
-  children?: ReactElement;
-  startDegree?: number;
-  endDegree?: number;
-  rotate?: number;
-};
-
-export type RectangleProps = {
-  backgroundColor?: string;
-  color?: string;
-  clockwise?: boolean;
-  speed?: number;
-  stroke?: number;
-  height?: number;
-  width?: number;
-  children?: ReactElement;
-  startDegree?: number;
-  endDegree?: number;
-  rotate?: number;
-};
